requirements.txt
asgiref==3.9.1
charset-normalizer==3.4.2
Django==5.2.4
et_xmlfile==2.0.0
lxml==6.0.0
openpyxl==3.1.5
pillow==11.3.0
python-docx==1.2.0
reportlab==4.4.2
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
